requirements.txt
gspread
oauth2client
psycopg2-binary
sqlalchemy
python-dotenv